refactor: MemberRepository 추상화
