+++ src/main/java/cart/application/CouponService.java
@@ -6,6 +6,7 @@ import cart.dto.CouponIssueRequest;
 import cart.dto.CouponReissueRequest;
 import cart.exception.CannotChangeCouponStatusException;
 import cart.repository.CouponRepository;
+import cart.repository.MemberJdbcRepository;
 import cart.repository.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +18,7 @@ public class CouponService {
     private final CouponRepository couponRepository;
     private final MemberRepository memberRepository;
 
-    public CouponService(final CouponRepository couponRepository, final MemberRepository memberRepository) {
+    public CouponService(final CouponRepository couponRepository, final MemberJdbcRepository memberRepository) {
         this.couponRepository = couponRepository;
         this.memberRepository = memberRepository;
     }
